Implement book restful endpoints

+++ app/Http/Controllers/BookController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\CreateBookRequest;
 use App\Http\Traits\RespondsWithHttpStatus;
 use App\Models\Book;
+use App\Rules\ValidIban;
 
 class BookController extends Controller
 {
@@ -11,12 +13,14 @@ class BookController extends Controller
 
     public function index()
     {
+        $books = Book::all();
 
+        return $this->success('Books Retrieved Successfully', $books);
     }
 
     public function show(Book $book)
     {
-        return $this->success('Book Retrieved Successfully', [$book]);
+        return $this->success('Book Retrieved Successfully', $book);
     }
 
     public function update(Book $book)
@@ -24,14 +28,22 @@ class BookController extends Controller
 
     }
 
-    public function store()
+    public function store(CreateBookRequest $createBookRequest)
     {
+        $this->validate($createBookRequest, ['isbn' => new ValidIban()]);
 
+        $validated = $createBookRequest->validated();
+
+        Book::create($validated);
+
+        return $this->success('Book Created Successfully');
     }
 
     public function delete(Book $book)
     {
-        $book->delete();
+        if ($book) {
+            $book->delete();
+        }
 
         return $this->success('Book Deleted Successfully');
     }
